requirements.txt
fastapi
uvicorn[standard]
pika
python-dotenv
numpy
